format modules structure
