+++ data/labels.json
@@ -6,5 +6,13 @@
   {
     "title": "new",
     "color": "red"
+  },
+  {
+    "title": "new",
+    "color": "blue"
+  },
+  {
+    "title": "new",
+    "color": "red"
   }
 ]
