+++ data/music_albums.json
@@ -48,5 +48,15 @@
       "id": 31,
       "name": "test"
     }
+  },
+  {
+    "published_date": "2009-09-09",
+    "label": "we",
+    "author": "wew",
+    "on_spotify": false,
+    "genre": {
+      "id": 59,
+      "name": "tt"
+    }
   }
 ]
